vpnkit-forwarder: remove dead code
In the past we used to act like busybox and switch on argv[0] between
- exposing a single port, a function now performed by vpnkit-expose-port
- handling incoming connections from the host and forwarding them to
  the backend

Ever since we started using a separate `vpnkit-expose-port`, this first
set of code has been dead so we remove it.
